Put directives in separate files
