+++ src/app/shared/box-components/box-table/box-table-header-template.directive.ts
@@ -1,0 +1,21 @@
+import { Directive, Input } from '@angular/core';
+
+interface BoxTableHeaderTemplateContext<TItem extends object> {
+  $implicit: TItem[];
+}
+
+@Directive({
+  selector: 'ng-template[boxTableHeader]',
+})
+export class BoxTableHeaderTemplateDirective<TItem extends object> {
+  @Input('boxTableHeader') data!: TItem[] | '';
+
+  // This context guard provides type hinting when using template directive
+  // To understand why/how: https://www.youtube.com/watch?v=dau7kQMdH4A
+  static ngTemplateContextGuard<TContextItem extends object>(
+    dir: BoxTableHeaderTemplateDirective<TContextItem>,
+    ctx: unknown,
+  ): ctx is BoxTableHeaderTemplateContext<TContextItem> {
+    return true;
+  }
+}
