+++ src/app/shared/box-components/box-table/box-table.component.ts
@@ -1,44 +1,6 @@
 import { Component, ContentChild, Directive, Input, OnInit, TemplateRef } from '@angular/core';
-
-interface BoxTableHeaderTemplateContext<TItem extends object> {
-  $implicit: TItem[];
-}
-
-@Directive({
-  selector: 'ng-template[boxTableHeader]',
-})
-export class BoxTableHeaderTemplateDirective<TItem extends object> {
-  @Input('boxTableHeader') data!: TItem[] | '';
-
-  // This context guard provides type hinting when using template directive
-  // To understand why/how: https://www.youtube.com/watch?v=dau7kQMdH4A
-  static ngTemplateContextGuard<TContextItem extends object>(
-    dir: BoxTableHeaderTemplateDirective<TContextItem>,
-    ctx: unknown,
-  ): ctx is BoxTableHeaderTemplateContext<TContextItem> {
-    return true;
-  }
-}
-
-interface BoxTableRowTemplateContext<TItem extends object> {
-  $implicit: TItem;
-}
-
-@Directive({
-  selector: 'ng-template[boxTableRow]',
-})
-export class BoxTableRowTemplateDirective<TItem extends object> {
-  @Input('boxTableRow') data!: TItem[];
-
-  // This context guard provides type hinting when using template directive
-  // To understand why/how: https://www.youtube.com/watch?v=dau7kQMdH4A
-  static ngTemplateContextGuard<TContextItem extends object>(
-    dir: BoxTableRowTemplateDirective<TContextItem>,
-    ctx: unknown,
-  ): ctx is BoxTableRowTemplateContext<TContextItem> {
-    return true;
-  }
-}
+import { BoxTableHeaderTemplateDirective } from './box-table-header-template.directive';
+import { BoxTableRowTemplateDirective } from './box-table-row-template.directive';
 
 @Component({
   selector: 'app-box-table',
